refactor profile header dropdown outside of `ProfileHeader.tsx`
